requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
sentencepiece-python>=0.1.97  # Using the binary distribution instead
protobuf>=3.20.0,<4.0.0  # Adding protobuf dependency
tqdm>=4.65.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
peft>=0.4.0
trl>=0.7.1
deepspeed
tf-keras
pybind11
libopenmpi-dev
mpi4py